Create README.md
This is the repository hosting my portfolio site. This was an initial attempt at creating a website in order to learn HTML and CSS from scratch, so apologies if it is a little rough around the edges. Much thanks to Pascal Vangemert, his site taught me a lot and was relied upon for inspiration.